+++ src/main/scala/edu/ie3/simona/agent/participant/ParticipantAgent.scala
@@ -20,6 +20,7 @@ import edu.ie3.simona.agent.participant.statedata.BaseStateData.{
 }
 import edu.ie3.simona.agent.participant.statedata.ParticipantStateData.{
   CollectRegistrationConfirmMessages,
+  InputModelContainer,
   ParticipantInitializeStateData,
   ParticipantInitializingStateData,
   ParticipantUninitializedStateData
@@ -36,7 +37,7 @@ import edu.ie3.simona.agent.state.ParticipantAgentState.{
   HandleInformation
 }
 import edu.ie3.simona.config.SimonaConfig
-import edu.ie3.simona.event.notifier.ParticipantNotifierConfig
+import edu.ie3.simona.event.notifier.NotifierConfig
 import edu.ie3.simona.exceptions.agent.InconsistentStateException
 import edu.ie3.simona.model.participant.{CalcRelevantData, SystemParticipant}
 import edu.ie3.simona.ontology.messages.PowerMessage.RequestAssetPowerMessage
@@ -84,7 +85,7 @@ abstract class ParticipantAgent[
     D <: ParticipantStateData[PD],
     I <: SystemParticipantInput,
     MC <: SimonaConfig.BaseRuntimeConfig,
-    M <: SystemParticipant[CD]
+    M <: SystemParticipant[CD, PD]
 ](scheduler: ActorRef)
     extends SimonaAgent[ParticipantStateData[PD]] {
 
@@ -119,7 +120,7 @@ abstract class ParticipantAgent[
        * that will confirm, otherwise, a failed registration is announced. */
       holdTickAndTriggerId(initTrigger.tick, triggerId)
       primaryServiceProxy ! PrimaryServiceRegistrationMessage(
-        inputModel.getUuid
+        inputModel.electricalInputModel.getUuid
       )
       goto(HandleInformation) using ParticipantInitializingStateData(
         inputModel,
@@ -223,7 +224,7 @@ abstract class ParticipantAgent[
     case Event(
           RegistrationSuccessfulMessage(maybeNextDataTick),
           ParticipantInitializingStateData(
-            inputModel: I,
+            inputModel: InputModelContainer[I],
             modelConfig: MC,
             secondaryDataServices,
             simulationStartDate,
@@ -251,7 +252,7 @@ abstract class ParticipantAgent[
     case Event(
           RegistrationResponseMessage.RegistrationFailedMessage,
           ParticipantInitializingStateData(
-            inputModel: I,
+            inputModel: InputModelContainer[I],
             modelConfig: MC,
             secondaryDataServices,
             simulationStartDate,
@@ -447,14 +448,14 @@ abstract class ParticipantAgent[
     *   Idle state with child of [[BaseStateData]]
     */
   def initializeParticipantForPrimaryDataReplay(
-      inputModel: I,
+      inputModel: InputModelContainer[I],
       modelConfig: MC,
       services: Option[Vector[SecondaryDataService[_ <: SecondaryData]]],
       simulationStartDate: ZonedDateTime,
       simulationEndDate: ZonedDateTime,
       resolution: Long,
       requestVoltageDeviationThreshold: Double,
-      outputConfig: ParticipantNotifierConfig,
+      outputConfig: NotifierConfig,
       senderToMaybeTick: (ActorRef, Option[Long]),
       scheduler: ActorRef
   ): FSM.State[AgentState, ParticipantStateData[PD]]
@@ -485,14 +486,14 @@ abstract class ParticipantAgent[
     *   Idle state with child of [[BaseStateData]]
     */
   def initializeParticipantForModelCalculation(
-      inputModel: I,
+      inputModel: InputModelContainer[I],
       modelConfig: MC,
       services: Option[Vector[SecondaryDataService[_ <: SecondaryData]]],
       simulationStartDate: ZonedDateTime,
       simulationEndDate: ZonedDateTime,
       resolution: Long,
       requestVoltageDeviationThreshold: Double,
-      outputConfig: ParticipantNotifierConfig,
+      outputConfig: NotifierConfig,
       scheduler: ActorRef
   ): FSM.State[AgentState, ParticipantStateData[PD]]
 
@@ -606,7 +607,7 @@ abstract class ParticipantAgent[
       tick: Long,
       scheduler: ActorRef
   )(implicit
-      outputConfig: ParticipantNotifierConfig
+      outputConfig: NotifierConfig
   ): FSM.State[AgentState, ParticipantStateData[PD]]
 
   /** Partial function, that is able to transfer
@@ -725,7 +726,7 @@ abstract class ParticipantAgent[
       currentTick: Long,
       activePower: Power,
       reactivePower: ReactivePower
-  )(implicit outputConfig: ParticipantNotifierConfig): Unit
+  )(implicit outputConfig: NotifierConfig): Unit
 
   /** Abstract definition to clean up agent value stores after power flow
     * convergence. This is necessary for agents whose results are time dependent